Divided Store Controller into DesnghTheme and PremiumSubscription controllers
